+++ src/main/java/com/semihbkgr/corbeau/controller/ApplicationController.java
@@ -35,23 +35,9 @@ public class ApplicationController {
 
     @GetMapping
     public Mono<String> menu(final Model model) {
-        var postDeepTagListCombinationsReactiveData=new ReactiveDataDriverContextVariable(
-                postService.findAllByActivatedDeep(
-                        true,
-                        PageRequest.of(0, POST_PAGE_SIZE, Sort.by("updated_at").descending())
-                        )
-                        .flatMapSequential(postDeep->
-                                tagService.findAllByPostId(postDeep.getId())
-                                        .collectList()
-                                        .map(list->new PostDeepTagList(postDeep,list)))
-                , 1);
-        model.addAttribute("postDeepTagListCombinations",postDeepTagListCombinationsReactiveData);
-        return tagService.findAllByActivatedDeep(true)
-                .collectList()
-                .flatMapMany(tagDeepList->{
-                    model.addAttribute("tags",tagDeepList);
-                    return subjectService.findAll();
-                })
+        var tagsReactiveData=new ReactiveDataDriverContextVariable(tagService.findAllDeep(), 1);
+        model.addAttribute("tags",tagsReactiveData);
+        return subjectService.findAll()
                 .collectList()
                 .map(subjectList -> {
                     model.addAttribute("subjects", subjectList);