+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Further Release
+## Further Release Features
 
 ---
 
@@ -6,3 +6,5 @@
 - Enable to add new moderator account
 - Check priveleges by moderator's role
 - Better search
+- Subject and tag order
+- Delete api